+++ packages-content-model/roosterjs-content-model/lib/createContentModelEditor.ts
@@ -4,11 +4,11 @@ import {
     ContentModelPastePlugin,
     EntityDelimiterPlugin,
 } from 'roosterjs-content-model-plugins';
-import type { EditorPlugin } from 'roosterjs-editor-types';
 import type {
     ContentModelEditorOptions,
     IContentModelEditor,
 } from 'roosterjs-content-model-editor';
+import type { EditorPlugin } from 'roosterjs-content-model-types';
 
 /**
  * Create a Content Model Editor using the given options
@@ -23,9 +23,8 @@ export function createContentModelEditor(
     additionalPlugins?: EditorPlugin[],
     initialContent?: string
 ): IContentModelEditor {
-    const legacyPlugins = additionalPlugins ? [...additionalPlugins] : [];
-    legacyPlugins.push(new ContentModelEditPlugin(), new EntityDelimiterPlugin());
-    const plugins = [new ContentModelPastePlugin()];
+    const legacyPlugins = [new ContentModelEditPlugin(), new EntityDelimiterPlugin()];
+    const plugins = [new ContentModelPastePlugin(), ...(additionalPlugins ?? [])];
 
     const options: ContentModelEditorOptions = {
         legacyPlugins: legacyPlugins,
