+++ packages-content-model/roosterjs-content-model/package.json
@@ -3,7 +3,6 @@
     "description": "Content Model for roosterjs (Under development)",
     "dependencies": {
         "tslib": "^2.3.1",
-        "roosterjs-editor-types": "",
         "roosterjs-content-model-types": "",
         "roosterjs-content-model-dom": "",
         "roosterjs-content-model-core": "",
